refactor: 주문(Order) 관련 클래스를 주문 패키지 아래에 추가.
