+++ src/main/java/kitchenpos/order/domain/enums/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain.enums;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
