+++ src/main/java/kitchenpos/order/repository/OrderRepository.java
@@ -1,4 +1,8 @@
-package kitchenpos.domain;
+package kitchenpos.order.repository;
+
+import kitchenpos.domain.OrderTable;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.enums.OrderStatus;
 
 import java.util.List;
 import java.util.Optional;
