+++ src/main/java/kitchenpos/order/service/OrderService.java
@@ -1,9 +1,14 @@
-package kitchenpos.application;
+package kitchenpos.order.service;
 
 import kitchenpos.domain.*;
 import kitchenpos.infra.KitchenridersClient;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.repository.MenuRepository;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.enums.OrderStatus;
+import kitchenpos.order.domain.enums.OrderType;
+import kitchenpos.order.repository.OrderRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
