+++ src/main/java/kitchenpos/order/web/OrderRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.ui;
+package kitchenpos.order.web;
 
-import kitchenpos.application.OrderService;
-import kitchenpos.domain.Order;
+import kitchenpos.order.service.OrderService;
+import kitchenpos.order.domain.Order;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
