+++ src/test/java/kitchenpos/application/InMemoryOrderRepository.java
@@ -1,8 +1,8 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderRepository;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.repository.OrderRepository;
+import kitchenpos.order.domain.enums.OrderStatus;
 import kitchenpos.domain.OrderTable;
 
 import java.util.*;
