+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -2,6 +2,12 @@ package kitchenpos.application;
 
 import kitchenpos.domain.*;
 import kitchenpos.menu.repository.MenuRepository;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.enums.OrderStatus;
+import kitchenpos.order.domain.enums.OrderType;
+import kitchenpos.order.repository.OrderRepository;
+import kitchenpos.order.service.OrderService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
